// packages/strapi-client/src/client/types.ts
export interface StrapiContentType {
  id: string;
  singularName: string;
  pluralName: string;
}
export interface StrapiClientOptions {
  url?: string;
  prefix?: string;
  contentTypes: string[];
  jwt?: string | null;
  axiosConfig?: AxiosRequestConfig;
}